refactor: sample 関数を src/utils 配下に移動
