--- src/tests/utils/sample/index.ts
@@ -1,3 +1,0 @@
-import { sample } from "./sample";
-
-export default sample;
